moved is_in_package to common
